# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AcronIQ Forge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  repoUrl: string;
  status: string;
  healthScore: number;
  language: string;
  branch: string;
  /** @nullable */
  lastScanAt?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  repoUrl: string;
  language?: string;
  branch?: string;
}

export interface ProjectUpdate {
  name?: string;
  repoUrl?: string;
  language?: string;
  branch?: string;
  status?: string;
}

export interface ProjectSummary {
  projectId: number;
  healthScore: number;
  openIssues: number;
  criticalIssues: number;
  /** @nullable */
  lastScanAt: string | null;
  filesIndexed: number;
  openTasks: number;
  openPullRequests: number;
}

export interface Issue {
  id: number;
  projectId: number;
  type: string;
  severity: string;
  filePath: string;
  message: string;
  /** @nullable */
  line?: number | null;
  status: string;
  createdAt: string;
}

export interface Task {
  id: number;
  projectId: number;
  projectName?: string;
  type: string;
  status: string;
  prompt: string;
  /** @nullable */
  plan?: string | null;
  filesModified?: string[];
  /** @nullable */
  buildStatus?: string | null;
  /** @nullable */
  confidenceScore?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface TaskInput {
  projectId: number;
  /** @minLength 1 */
  prompt: string;
  type?: string;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  projectId: number | null;
  /** @nullable */
  projectName?: string | null;
  messageCount: number;
  createdAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  projectId?: number | null;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface PullRequest {
  id: number;
  projectId: number;
  projectName?: string;
  /** @nullable */
  taskId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  branch: string;
  /** @nullable */
  url?: string | null;
  filesChanged?: number;
  additions?: number;
  deletions?: number;
  createdAt: string;
  /** @nullable */
  mergedAt?: string | null;
}

export interface AuditEntry {
  id: number;
  entityType: string;
  /** @nullable */
  entityId?: number | null;
  action: string;
  actor: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  openTasks: number;
  awaitingApproval: number;
  openPullRequests: number;
  criticalIssues: number;
  avgHealthScore: number;
  tasksThisWeek: number;
  connectedDevices: number;
  agentsEnabled: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  projectName: string;
  /** @nullable */
  entityId?: number | null;
  /** @nullable */
  severity?: string | null;
  createdAt: string;
}

export interface Device {
  id: number;
  userId: string;
  name: string;
  /** windows | macos | linux */
  platform: string;
  /** online | offline | idle */
  status: string;
  pairingToken: string;
  /** @nullable */
  agentVersion?: string | null;
  ollamaAvailable: boolean;
  /** @nullable */
  ollamaVersion?: string | null;
  /** @nullable */
  activeModel?: string | null;
  /** @nullable */
  lastHeartbeatAt?: string | null;
  createdAt: string;
}

export interface DeviceInput {
  /** @minLength 1 */
  name: string;
  platform?: string;
}

export interface DeviceHeartbeat {
  status?: string;
  ollamaAvailable?: boolean;
  /** @nullable */
  ollamaVersion?: string | null;
  /** @nullable */
  activeModel?: string | null;
  /** @nullable */
  agentVersion?: string | null;
}

export interface Agent {
  id: number;
  userId: string;
  /** repair | architecture | security | documentation */
  type: string;
  name: string;
  description: string;
  enabled: boolean;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  lastRunAt?: string | null;
  createdAt: string;
}

export interface AgentInput {
  type: string;
  /** @minLength 1 */
  name: string;
  description: string;
  /** @nullable */
  projectId?: number | null;
}

export interface AgentUpdate {
  enabled?: boolean;
  /** @nullable */
  projectId?: number | null;
}

export interface AgentRun {
  id: number;
  agentId: number;
  projectId: number;
  /** running | completed | failed */
  status: string;
  /** @nullable */
  summary?: string | null;
  recommendations?: string[];
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface AgentRunInput {
  projectId: number;
}

export interface MemoryEntry {
  id: number;
  userId: string;
  /** @nullable */
  projectId?: number | null;
  /** architecture | fix | convention | preference | workflow */
  category: string;
  key: string;
  value: string;
  /** manual | agent | scan */
  source: string;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryEntryInput {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  key: string;
  /** @minLength 1 */
  value: string;
  /** @nullable */
  projectId?: number | null;
  source?: string;
}

export interface MemoryEntryUpdate {
  value?: string;
  category?: string;
}

export interface ConstitutionRule {
  id: number;
  userId: string;
  /** language | security | structure | git | testing | architecture */
  category: string;
  title: string;
  description: string;
  /** block | warn | info */
  enforcement: string;
  enabled: boolean;
  createdAt: string;
}

export interface ConstitutionRuleInput {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  enforcement?: string;
}

export interface ConstitutionRuleUpdate {
  enabled?: boolean;
  enforcement?: string;
  title?: string;
  description?: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListTasksParams = {
/**
 * @nullable
 */
projectId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};

export type ListPullRequestsParams = {
/**
 * @nullable
 */
status?: string | null;
};

export type ListAuditEntriesParams = {
/**
 * @nullable
 */
entityType?: string | null;
};

export type ListMemoryEntriesParams = {
/**
 * @nullable
 */
projectId?: number | null;
/**
 * @nullable
 */
category?: string | null;
};