# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcronIQ Forge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /devices/{id}/heartbeat:
    patch:
      operationId: deviceHeartbeat
      tags: [devices]
      summary: Device sends heartbeat with status update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceHeartbeat"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all configured agents
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent (enable/disable, set project)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}/run:
    post:
      operationId: runAgent
      tags: [agents]
      summary: Trigger an agent run against a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunInput"
      responses:
        "202":
          description: Run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{id}/runs:
    get:
      operationId: listAgentRuns
      tags: [agents]
      summary: List recent runs for an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
...
  /constitution:
    get:
      operationId: listConstitutionRules
      tags: [constitution]
      summary: List engineering constitution rules
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConstitutionRule"
    post:
      operationId: createConstitutionRule
      tags: [constitution]
      summary: Add a constitution rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstitutionRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstitutionRule"
  /constitution/{id}:
    patch:
      operationId: updateConstitutionRule
      tags: [constitution]
      summary: Update a rule (enable/disable, change enforcement)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConstitutionRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstitutionRule"
    delete:
      operationId: deleteConstitutionRule
      tags: [constitution]
      summary: Delete a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Device:
      type: object
      required: [id, userId, name, platform, status, pairingToken, createdAt, ollamaAvailable]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        platform:
          type: string
          description: "windows | macos | linux"
        status:
          type: string
          description: "online | offline | idle"
        pairingToken:
          type: string
        agentVersion:
          type: ["string", "null"]
        ollamaAvailable:
          type: boolean
        ollamaVersion:
          type: ["string", "null"]
        activeModel:
          type: ["string", "null"]
        lastHeartbeatAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DeviceHeartbeat:
      type: object
      properties:
        status:
          type: string
        ollamaAvailable:
          type: boolean
        ollamaVersion:
          type: ["string", "null"]
        activeModel:
          type: ["string", "null"]
        agentVersion:
          type: ["string", "null"]
    Agent:
      type: object
      required: [id, userId, type, name, description, enabled, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          description: "repair | architecture | security | documentation"
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        projectId:
          type: ["integer", "null"]
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AgentInput:
      type: object
      required: [type, name, description]
      properties:
        type:
          type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        projectId:
          type: ["integer", "null"]
    AgentUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        projectId:
          type: ["integer", "null"]
    AgentRun:
      type: object
      required: [id, agentId, projectId, status, createdAt]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          description: "running | completed | failed"
        summary:
          type: ["string", "null"]
        recommendations:
          type: array
          items:
            type: string
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    AgentRunInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
...
    ConstitutionRule:
      type: object
      required: [id, userId, category, title, description, enforcement, enabled, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        category:
          type: string
          description: "language | security | structure | git | testing | architecture"
        title:
          type: string
        description:
          type: string
        enforcement:
          type: string
          description: "block | warn | info"
        enabled:
          type: boolean
        createdAt:
          type: string
    ConstitutionRuleInput:
      type: object
      required: [category, title, description]
      properties:
        category:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        enforcement:
          type: string
    ConstitutionRuleUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        enforcement:
          type: string
        title:
          type: string
        description:
          type: string